Blow away non-deterministic streaming dataset class, replace with resumable rewrite.
