+++ streaming/__init__.py
@@ -7,9 +7,9 @@ import streaming.text as text
 import streaming.vision as vision
 from streaming._version import __version__
 from streaming.base import (CSVWriter, Dataset, JSONWriter, LocalIterableDataset, LocalMapDataset,
-                            LocalResumableDataset, MDSWriter, TSVWriter, XSVWriter)
+                            MDSWriter, TSVWriter, XSVWriter)
 
 __all__ = [
     'Dataset', 'CSVWriter', 'JSONWriter', 'MDSWriter', 'TSVWriter', 'XSVWriter',
-    'LocalIterableDataset', 'LocalMapDataset', 'LocalResumableDataset', 'vision', 'text'
+    'LocalIterableDataset', 'LocalMapDataset', 'vision', 'text'
 ]
