+++ streaming/base/__init__.py
@@ -3,11 +3,10 @@
 
 """MosaicML Streaming Datasets for cloud-native model training."""
 
-from streaming.base.dataset import (Dataset, LocalIterableDataset, LocalMapDataset,
-                                    LocalResumableDataset)
+from streaming.base.dataset import Dataset, LocalIterableDataset, LocalMapDataset
 from streaming.base.format import CSVWriter, JSONWriter, MDSWriter, TSVWriter, XSVWriter
 
 __all__ = [
-    'Dataset', 'CSVWriter', 'JSONWriter', 'LocalIterableDataset', 'LocalMapDataset',
-    'LocalResumableDataset', 'MDSWriter', 'TSVWriter', 'XSVWriter'
+    'Dataset', 'CSVWriter', 'JSONWriter', 'LocalIterableDataset', 'LocalMapDataset', 'MDSWriter',
+    'TSVWriter', 'XSVWriter'
 ]
