+++ streaming/base/dataset/streaming.py
@@ -1,38 +1,65 @@
 # Copyright 2022 MosaicML Streaming authors
 # SPDX-License-Identifier: Apache-2.0
 
-"""The :class:`Dataset` class, used for building streaming iterable datasets."""
+"""A streaming pytorch IterableDataset, resumable mid-epoch, whose shards reside locally."""
 
 import json
 import os
 from enum import IntEnum
-from multiprocessing import Pool
-from threading import RLock, Thread
+from multiprocessing.shared_memory import SharedMemory
+from threading import Thread
 from time import sleep
-from typing import Any, Iterator, List, Optional
+from typing import Any, Dict, Iterator, List, Optional, Tuple
 
 import numpy as np
+import torch
+from filelock import FileLock
+from numpy.typing import NDArray
+from torch import distributed as dist
 from torch.utils.data import IterableDataset
 
-from streaming.base import distributed as dist
-from streaming.base.compression import decompress
-from streaming.base.download import download_or_wait, wait_for_download
+from streaming.base.download import download
 from streaming.base.format import reader_from_json
 from streaming.base.format.base.reader import FileInfo
 from streaming.base.hashing import get_hash
-from streaming.base.index import Index, Partition, get_index_basename
+from streaming.base.index import Index
+from streaming.base.shared import SharedBarrier
+from streaming.base.shuffle import get_epoch
+from streaming.base.world import World
 
 
-class DownloadStatus(IntEnum):
-    """Download status."""
-    NOT_STARTED = 1
-    IN_PROGRESS = 2
-    DONE = 3
-    FAILED = 4
+class _ShardState(IntEnum):
+    """The download status of a shard.
+
+    Restrictions:
+    - The initial state of UNKNOWN must be zero.
+    - The state will only ever change in the upward direction.
+    """
+    UNKNOWN = 0
+    DOWNLOADING = 1
+    DOWNLOADED = 2
 
 
 class Dataset(IterableDataset):
-    """A sharded, streamed, iterable dataset.
+    """A streaming pytorch IterableDataset, resumable mid-epoch, whose shards reside locally.
+
+    Training is represented as sequence of one or more training sessions, which are cleared between
+    epochs. A training session is an array of how many samples each worker has processed during
+    this session.
+
+    To restore from checkpoint, even while changing the number of worker partitions, we recreate
+    the deterministic initial shuffle then replay the training history: splitting, truncating from
+    front, and rejoining for each session in order.
+
+    We communicate this state across all ranks and worker processes by putting it in shared memory
+    objects which are updated during checkpointing and training.
+
+    Checkpoints are represented in JSON as follows:
+
+        {
+            'epoch': int,
+            'sessions': List[List[int]],
+        }
 
     Args:
         local (str): Local dataset directory where shards are cached by split.
@@ -40,324 +67,267 @@ class Dataset(IterableDataset):
             rank and worker's partition of the dataset must all exist locally. Defaults to
             ``None``.
         split (str, optional): Which dataset split to use, if any. Defaults to ``None``.
-        shuffle (bool): Whether to shuffle the samples while iterating. Defaults to ``True``.
-        prefetch (int, optional): Target number of samples remaining to prefetch while iterating.
-            Defaults to ``None``.
+        shuffle (bool): Whether to shuffle the samples while iterating. Defaults to ``False``.
+        predownload (int, optional): Target number of samples remaining to prefetch while
+            iterating. Defaults to ``None``.
         keep_zip (bool, optional): Whether to keep or delete the compressed file when
             decompressing downloaded shards. If set to None, keep iff remote is local. Defaults to
             ``None``.
-        retry (int): Number of download re-attempts before giving up. Defaults to ``2``.
-        timeout (float): Number of seconds to wait for a shard to download before raising an
-            exception. Defaults to ``60``.
-        hash (str, optional): Optional hash or checksum algorithm to use to validate shards.
-            Defaults to ``None``.
+        download_retry (int): Number of download re-attempts before giving up. Defaults to ``2``.
+        download_timeout (float): Number of seconds to wait for a shard to download before raising
+            an exception. Defaults to ``60``.
+        validate_hash (str, optional): Optional hash or checksum algorithm to use to validate
+            shards. Defaults to ``None``.
         batch_size (int, optional): Hint the batch_size that will be used on each device's
             DataLoader. Defaults to ``None``.
-
-    .. doctest::
-
-        To write the dataset:
-        >>> import numpy as np
-        >>> from PIL import Image
-        >>> from uuid import uuid4
-        >>> from streaming import MDSWriter
-        >>> dirname = 'dirname'
-        >>> columns = {
-        ...     'uuid': 'str',
-        ...     'img': 'jpeg',
-        ...     'clf': 'int'
-        ... }
-        >>> compression = 'zstd'
-        >>> hashes = 'sha1', 'xxh64'
-        >>> samples = [
-        ...     {
-        ...         'uuid': str(uuid4()),
-        ...         'img': Image.fromarray(np.random.randint(0, 256, (32, 48, 3), np.uint8)),
-        ...         'clf': np.random.randint(10),
-        ...     }
-        ...     for i in range(1000)
-        ... ]
-        >>> with MDSWriter(dirname, columns, compression, hashes) as out:
-        ...     for sample in samples:
-        ...         out.write(sample)
-
-        To read the dataset:
-        >>> from streaming import Dataset
-        >>> dataset = Dataset(dirname)
-        >>> for sample in dataset:
-        ...     print(sample)
-
-        To read the dataset (with all optional arguments):
-        >>> from streaming import Dataset
-        >>> dataset = Dataset(local=dirname, remote=None, split=None, shuffle=True,
-        ...                   prefetch=100_000, keep_zip=None, retry=2, timeout=60, hash=None,
-        ...                   batch_size=None)
+        seed (int, optional): Seed for shuffling, or ``None`` for random seed. Defaults to
+            ``None``.
+        batch_size (int, optional): Batch size of its DataLoader, which affects how the dataset is
+            partitioned over the workers. Defaults to ``None``.
+        num_workers (int, optional): Number of workers of its DataLoader, which determines the size
+            of the barrier to coordinate workers while iterating. Defaults to ``None``.
     """
 
     def __init__(self,
                  local: str,
                  remote: Optional[str] = None,
                  split: Optional[str] = None,
-                 shuffle: bool = True,
-                 prefetch: Optional[int] = 100_000,
+                 shuffle: bool = False,
+                 predownload: Optional[int] = 100_000,
                  keep_zip: Optional[bool] = None,
-                 retry: int = 2,
-                 timeout: float = 60,
-                 hash: Optional[str] = None,
-                 batch_size: Optional[int] = None) -> None:
-        if keep_zip is None:
-            keep_zip = remote is None or remote == local
-        hash = hash or None
-
+                 download_retry: int = 2,
+                 download_timeout: float = 60,
+                 validate_hash: Optional[str] = None,
+                 seed: Optional[int] = None,
+                 batch_size: Optional[int] = None,
+                 num_workers: Optional[int] = None):
         self.local = local
         self.remote = remote
-        self.split = split or ''
+        self.split = split or ''  # Empty string for os.path.join().
         self.shuffle = shuffle
-        self.prefetch = prefetch
+        self.predownload = predownload
         self.keep_zip = keep_zip
-        self.retry = retry
-        self.timeout = timeout
-        self.hash = hash
+        self.download_retry = download_retry
+        self.download_timeout = download_timeout
+        self.validate_hash = validate_hash or None
+        # Seed is set below.
         self.batch_size = batch_size
+        self.num_workers = num_workers
 
-        basename = get_index_basename()
-        wait = dist.get_local_rank() != 0
-        filename = self._download_file(basename, wait)
+        # Load the index.json file.
+        filename = os.path.join(local, split, 'index.json')  # pyright: ignore
         obj = json.load(open(filename))
         assert obj['version'] == 2
 
+        # Initialize shard readers according to the loaded info.
         self.shards = []
         for info in obj['shards']:
             shard = reader_from_json(local, split, info)
             self.shards.append(shard)
 
-        samples_per_shard = np.array([x.samples for x in self.shards])
-        self.index = Index(samples_per_shard, batch_size)
+        # Build the Index (for partitioning and mapping samples to shards).
+        self.shard_sizes = np.array([x.samples for x in self.shards])
+        self.index = Index(self.shard_sizes)
 
-        # Fields, protected by the lock, relating to loading shards in the background.
-        self._lock: RLock
-        self._has_shard = np.zeros(len(self.shards), np.uint8)
-        self._next_epoch = 0
-        self._epoch_to_todo_ids = {}
-        self._downloaded_ids = []
-        self._download_status = DownloadStatus.NOT_STARTED
-        self._download_exception: Exception
+        # Setup for coordinating.
+        world = World()
+        device = torch.device(f'cuda:{world.rank_of_node}')
+        tensor = torch.zeros(1, dtype=torch.int64, device=device)
+
+        # Coordinate the seed across ranks.
+        if world.is_leader:
+            if seed is None:
+                seed = np.random.randint(1 << 60)
+            tensor[0] = seed
+        dist.broadcast(tensor, 0)
+        self.seed = int(tensor)
+
+        # Add a coordinated random prefix to all shm names for uniqueness.
+        if world.is_leader:
+            tensor[0] = np.random.randint(1 << 60)
+        dist.broadcast(tensor, 0)
+        self._prefix = f'{int(tensor):016x}_{self.split}'
+
+        # Set up the epoch counter.
+        #
+        # Note: we do not assume that the end of __iter__() will ever be reached, so we need to
+        # increment the epoch counter at the start of __iter__() instead of at the end, so we need
+        # to track what the next epoch is, not the current epoch.
+        name = f'{self._prefix}_next_epoch'
+        size = np.int64().nbytes
+        try:
+            self._next_epoch_shm = SharedMemory(name, True, size)
+        except:
+            self._next_epoch_shm = SharedMemory(name)
+        self._next_epoch_arr = np.ndarray(1, buffer=self._next_epoch_shm.buf, dtype=np.int64)
+        self._next_epoch_arr[0] = 0
+
+        # Placeholder for _resume_shm, a shared memory object where load_state_dict() saves its
+        # data to be picked up by __iter__().
+        self._resume_shm = None
+
+        # Create the barrier.
+        total_workers = world.ranks_per_node * (self.num_workers or 1)
+        self._barrier_filelock_path = os.path.join('/tmp', 'mds', self._prefix, 'barrier_filelock')
+        self._barrier_shm_path = f'{self._prefix}_barrier_shm'
+        self._barrier = SharedBarrier(total_workers, self._barrier_filelock_path,
+                                      self._barrier_shm_path)
+
+        self._iter_index = 0
+        self._download_index = 0
+
+    @property
+    def next_epoch(self) -> int:
+        """Get property next_epoch.
+
+        Returns:
+            int: Next epoch.
+        """
+        return int(self._next_epoch_arr[0])
+
+    @next_epoch.setter
+    def next_epoch(self, next_epoch: int) -> None:
+        """Set property next_epoch.
+
+        Args:
+            next_epoch (int): Next epoch.
+        """
+        self._next_epoch_arr[0] = next_epoch
 
     def __len__(self) -> int:
-        """Get the length as an IterableDataset (ie, divided by num devices).
+        """Get the length as an IterableDataset.
 
         Returns:
             int: Dataset length.
         """
         return self.index.get_samples_per_device()
 
-    def _load_shards(self, shards: List[int], partition: Partition) -> None:
-        """Load our partition's samples from the given locally cached shards.
-
-        Every time you call __iter__ on this dataset, it registers the list of samples you have
-        left, which will not be the full epoch if the dataset isn't finished loaded when you start
-        training.
-
-        Calls to this method during training modify the samples remaining on these iterations on
-        the fly to insert these new samples and then re-sort, making the shuffle as perfect as was
-        possible.
-
-        This operation is heavy and takes the lock, so call this method with all available shards
-        at once.
+    def __getitem__(self, index: int) -> Dict[str, Any]:
+        """Get sample by global index.
 
         Args:
-            shards (List[int]): Shard IDs.
-            partition (Partition): Our rank and worker's partition of the dataset.
-        """
-        # Get our partition and shards' sample ranges.
-        new_ids = []
-        for shard in shards:
-            shard_min_id = max(self.index.shard_offsets[shard], partition.min_sample_id)
-            shard_max_id = min(self.index.shard_offsets[shard + 1] - 1, partition.max_sample_id)
-            new_ids += list(range(shard_min_id, shard_max_id + 1))
-
-        with self._lock:
-            # Extend and optionally reshuffle the remaining samples of any epochs in progress.
-            if self.shuffle:
-                if self._download_status == DownloadStatus.IN_PROGRESS:
-                    self._downloaded_ids.extend(new_ids)
-                    np.random.shuffle(self._downloaded_ids)
-                for todo_ids in self._epoch_to_todo_ids.values():
-                    todo_ids.extend(new_ids)
-                    np.random.shuffle(todo_ids)
-            else:
-                if self._download_status == DownloadStatus.IN_PROGRESS:
-                    self._downloaded_ids.extend(new_ids)
-                for todo_ids in self._epoch_to_todo_ids.values():
-                    todo_ids.reverse()
-                    todo_ids.extend(new_ids)
-                    todo_ids.reverse()
-
-            # Note that we have loaded the shards.
-            for shard in shards:
-                self._has_shard[shard] = True
-
-    def _load_shard(self, shard: int, partition: Partition) -> None:
-        """Load our partition's samples from the given locally cached shard.
-
-        For performance reasons, prefer _load_shards() where possible.
-
-        Args:
-            shard (int): Shard ID.
-            partition (Partition): Our rank and worker's partition of the dataset.
-        """
-        self._load_shards([shard], partition)
-
-    def _decompress_shard_part(self, zip_info: FileInfo, zip_filename: str, raw_filename: str,
-                               compression: Optional[str], wait: bool):
-        """Validate and decompress shard data.
-
-        Args:
-            zip_info (FileInfo): Compressed file info.
-            zip_filename (str): Compressed filename.
-            raw_filename (str): Decompressed filename.
-            compression (str, optional): Compression algorithm.
-            wait (bool): Whether to wait for another worker to do the work.
-        """
-        # Doer or waiter?
-        if not wait:
-            # If doer: load compressed, validate, decompress, save decompressed, maybe remove
-            # compressed to save space.
-            data = open(zip_filename, 'rb').read()
-            if self.hash:
-                assert get_hash(self.hash, data) == zip_info.hashes[self.hash]
-            data = decompress(compression, data)  # pyright: ignore
-            tmp_filename = raw_filename + '.tmp'
-            with open(tmp_filename, 'wb') as out:
-                out.write(data)
-            os.rename(tmp_filename, raw_filename)
-            if not self.keep_zip:
-                os.remove(zip_filename)
-        else:
-            # If waiter: wait on decompressed to be written.
-            wait_for_download(raw_filename, self.timeout)
-
-    def _preload_shard_part(self,
-                            shard: int,
-                            partition: Partition,
-                            raw_info: FileInfo,
-                            zip_info: Optional[FileInfo] = None,
-                            compression: Optional[str] = None) -> bool:
-        """Decompress and validate shard data given raw/zip version metadata.
-
-        MDS format uses joint shards (ie, one file per shard). Other formats supported by streaming
-        use split shards (ie, shard data lives in two files per shard: the raw data itself and
-        metadata in a separate file).
-
-        Args:
-            shard (int): Shard ID.
-            partition (Partition): Our rank and worker's partition of the dataset.
-            raw_info (FileInfo): Raw file info.
-            zip_info (FileInfo, optional): Zip file info. Defaults to ``None``.
-            compression (str, optional): Compression algorithm used for zip_info. Defaults to
-                ``None``.
+            index (int): Sample index.
 
         Returns:
-            bool: Whether shard is present.
+            Dict[str, Any]: Column name with sample data.
         """
-        # Determine if we are the doer (process performing the preload), or the waiter (one waiting
-        # for it to complete). Multiple processes share a shard when it is split between workers.
-        wait = shard not in partition.shards_to_download
-
-        # If the decompressed version already exists, validate if requested.
-        raw_filename = os.path.join(self.local, self.split, raw_info.basename)
-        if os.path.isfile(raw_filename):
-            if not wait:
-                if self.hash:
-                    data = open(raw_filename, 'rb').read()
-                    assert get_hash(self.hash, data) == raw_info.hashes[self.hash]
-            return True
-
-        # If no decompressed file, and compression was not used, shard is missing.
-        if not zip_info:
-            return False
-
-        # If no decompressed file, compression was used, and no compressed file, shard is missing.
-        zip_filename = os.path.join(self.local, self.split, zip_info.basename)
-        if not os.path.isfile(zip_filename):
-            return False
-
-        # If has only the compressed version, decompress and maybe validate it.
-        self._decompress_shard_part(zip_info, zip_filename, raw_filename, compression, wait)
-        return True
-
-    def _preload_shard(self, shard: int, partition: Partition) -> bool:
-        """Decompress and validate a single shard, returning whether present.
-
-        Args:
-            shard (int): Which shard.
-
-        Returns:
-            bool: Whether shard is present.
-        """
-        assert shard in partition.shards
+        shard, index_in_shard = self.index.find_sample(index)
         reader = self.shards[shard]
-        for raw_info, zip_info in reader.file_pairs:
-            if not self._preload_shard_part(shard, partition, raw_info, zip_info,
-                                            reader.compression):
-                return False
-        return True
+        return reader[index_in_shard]
 
-    def _preload(self, partition: Partition) -> List[int]:
-        """Load any shards that are cached locally, returning missing shards.
+    def _resume(self, epoch: int) -> Tuple[int, List[NDArray[np.int64]]]:
+        """Resume from checkpoint.
 
         Args:
-            partition (Partition): Our rank and worker's partition of the dataset.
+            epoch (int): What epoch we think it is (pre-checkpoint).
 
         Returns:
-            List[int]: Missing shards that must be downloaded.
+            Tuple[int, List[NDArray[np.int64]]]: Pair of (resumed epoch, old sessions).
         """
-        # Create lock in preload() because we are prevented from putting it in __init__ because of
-        # DataLoader num_workers and fork/spawn semantics.
-        if not hasattr(self, '_lock'):
-            self._lock = RLock()
+        world = World()
 
-        # Bail out if has already been called.
-        with self._lock:
-            if self._download_status != DownloadStatus.NOT_STARTED:
-                return []
-            self._download_status = DownloadStatus.IN_PROGRESS
+        # Get the resume state, if it exists.
+        name = f'{self._prefix}_resume'
+        try:
+            shm = SharedMemory(name)
+        except:
+            # There is nothing to resume.
+            return epoch, []
 
-        # Find and load cached shards given our sample range.
-        present_shards = []
-        missing_shards = []
-        for shard in partition.shards:
-            if self._preload_shard(shard, partition):
-                present_shards.append(shard)
-            else:
-                missing_shards.append(shard)
-        self._load_shards(present_shards, partition)
+        # Parse the existent resume state.
+        obj = json.loads(bytes(shm.buf).decode('utf-8'))
 
-        # If there are no missing shards, we're done.
-        if not missing_shards:
-            with self._lock:
-                self._download_status = DownloadStatus.DONE
-            return []
+        # Check if the resume state is stale.
+        if obj['epoch'] < epoch:
+            # Clean up stale state.
+            if world.is_local_leader:
+                shm.unlink()
+            return epoch, []
 
-        # Always download the first shard first, if it is missing, because other workers may be
-        # waiting on it.
-        if self.shuffle:
-            if missing_shards[0] == partition.shards[0]:
-                nonfirst = 1
-            else:
-                nonfirst = 0
-            missing_shards = np.array(missing_shards)
-            np.random.shuffle(missing_shards[nonfirst:])
-            missing_shards = missing_shards.tolist()
+        # Load the correct epoch and previous training sessions this epoch.
+        epoch = obj['epoch']
+        old_sessions = [np.array(x) for x in obj['sessions']]
+        return epoch, old_sessions
 
-        return missing_shards
+    def _create_cur_session(self, epoch: int) -> Tuple[NDArray[np.int64], SharedMemory]:
+        """Create the current session, as we have just started an epoch.
 
-    def _download_file(self, basename: str, wait: bool = False) -> str:
+        Called by __iter__() in a worker process, or a per-rank process if workers aren't used.
+
+        Note: this also returns the underlying shared memory object, because the returned array
+        will become invalidated when it goes out of scope.
+
+        Args:
+            epoch (int): The current epoch.
+
+        Returns:
+            Tuple[NDArray[np.int64], SharedMemory]: Session and handle to shared memory.
+        """
+        world = World()
+        shm_name = f'{self._prefix}_session_{epoch}'
+        shm_bytes = world.num_workers * np.int64().nbytes
+        try:
+            shm = SharedMemory(shm_name, True, shm_bytes)
+        except:
+            shm = SharedMemory(shm_name)
+        cur_session = np.ndarray(world.num_workers, buffer=shm.buf, dtype=np.int64)
+        cur_session[:] = 0
+        return cur_session, shm
+
+    def _lookup_cur_session(
+            self, epoch: int) -> Tuple[Optional[NDArray[np.int64]], Optional[SharedMemory]]:
+        """Look up the current session, which exists if we are currently training.
+
+        Called by state_dict() in a per-rank process.
+
+        Note: this also returns the underlying shared memory object (if it exists), because the
+        returned array will become invalidated when it goes out of scope.
+
+        Args:
+            epoch (int): The current epoch.
+
+        Returns:
+            Tuple[Optional[NDArray[np.int64]], Optional[SharedMemory]]: Maybe session, maybe shm.
+        """
+        shm_name = f'{self._prefix}_session_{epoch}'
+        try:
+            shm = SharedMemory(shm_name)
+        except:
+            return None, None
+        num_workers = shm.size // np.int64().nbytes
+        cur_session = np.ndarray(num_workers, buffer=shm.buf, dtype=np.int64)
+        return cur_session, shm
+
+    def _get_partition(self, epoch: int, sessions: List[NDArray[np.int64]],
+                       world: World) -> NDArray[np.int64]:
+        # Local leader generates the partitions.
+        if world.is_local_leader:
+            sequences = get_epoch(self.shard_sizes, self.shuffle, self.seed, epoch, sessions)
+            base = world.node * world.ranks_per_node * world.workers_per_rank
+            for rank_of_node in range(world.ranks_per_node):
+                for worker_of_rank in range(world.workers_per_rank):
+                    worker = base + rank_of_node * world.workers_per_rank + worker_of_rank
+                    name = f'{self._prefix}_part_{worker:03}'
+                    sequence = sequences[worker]
+                    size = len(sequence) * np.int64(0).nbytes
+                    shm = SharedMemory(name, True, size)
+                    shm.buf[:] = sequence.tobytes()
+
+        self._barrier()
+
+        # Load our partition.
+        name = f'{self._prefix}_part_{world.worker:03}'
+        shm = SharedMemory(name)
+        todos = np.frombuffer(shm.buf, np.int64).copy()
+        shm.unlink()
+
+        return todos
+
+    def _download_file(self, basename: str) -> str:
         """Safely download a file from remote to local cache.
 
         Args:
             basename (str): Basename of file to download.
-            wait (bool): Whether to wait for another worker to download the file.
 
         Returns:
             str: Local cache filename.
@@ -367,12 +337,44 @@ class Dataset(IterableDataset):
         else:
             remote = os.path.join(self.remote, self.split, basename)
         local = os.path.join(self.local, self.split, basename)
-        download_or_wait(remote, local, wait, self.retry, self.timeout)
+        for _ in range(1 + self.download_retry):
+            try:
+                download(remote, local, self.download_timeout)
+            except:
+                continue
+            break
         return local
+
+    def _decompress_shard_part(self, zip_info: FileInfo, zip_filename: str, raw_filename: str,
+                               compression: Optional[str]) -> None:
+        """Validate and decompress shard data.
+
+        Args:
+            zip_info (FileInfo): Compressed file info.
+            zip_filename (str): Compressed filename.
+            raw_filename (str): Decompressed filename.
+            compression (str, optional): Compression algorithm.
+        """
+        # Load compressed.
+        data = open(zip_filename, 'rb').read()
+
+        # Validate what was downloaded.
+        if self.validate_hash:
+            assert get_hash(self.validate_hash, data) == zip_info.hashes[self.validate_hash]
+
+        # Decompress and save that.
+        data = decompress(compression, data)  # pyright: ignore
+        tmp_filename = raw_filename + '.tmp'
+        with open(tmp_filename, 'wb') as out:
+            out.write(data)
+        os.rename(tmp_filename, raw_filename)
+
+        # Maybe remove compressed to save space.
+        if not self.keep_zip:
+            os.remove(zip_filename)
 
     def _download_shard_part(self,
                              shard: int,
-                             partition: Partition,
                              raw_info: FileInfo,
                              zip_info: Optional[FileInfo] = None,
                              compression: Optional[str] = None) -> None:
@@ -384,7 +386,6 @@ class Dataset(IterableDataset):
 
         Args:
             shard (int): Shard ID.
-            partition (Partition): Our rank and worker's partition of the dataset.
             raw_info (FileInfo): Raw file info.
             zip_info (FileInfo, optional): Zip file info. Defaults to ``None``.
             compression (str, optional): Compression algorithm used for zip_info. Defaults to
@@ -395,238 +396,230 @@ class Dataset(IterableDataset):
         if os.path.isfile(raw_filename):
             return
 
-        # Determine if we are the doer (process performing the download), or the waiter (one waiting
-        # for it to complete). Multiple processes share a shard when it is split between workers.
-        wait = shard not in partition.shards_to_download
-
         # Is compression used?
         if zip_info:
-            # Download the compressed form if missing (or wait on its download).
+            # Download the compressed form if missing.
             zip_filename = os.path.join(self.local, self.split, zip_info.basename)
             if not os.path.isfile(zip_filename):
-                # Waiter or doer?
-                if wait:
-                    # If waiter: wait for *raw* version to exist (as the zip may be ephemeral).
-                    wait_for_download(raw_filename, self.timeout)
-                else:
-                    # If doer: download the zip version.
-                    self._download_file(zip_info.basename)
+                self._download_file(zip_info.basename)
 
-            # Validate and decompress (or wait on that).
-            self._decompress_shard_part(zip_info, zip_filename, raw_filename, compression, wait)
+            # Validate and decompress.
+            self._decompress_shard_part(zip_info, zip_filename, raw_filename, compression)
         else:
-            # Download the raw form (or wait on its download).
-            self._download_file(raw_info.basename, wait)
+            # Download the raw form.
+            self._download_file(raw_info.basename)
 
-            # Waiter or doer?
-            if not wait:
-                # If doer: validate if requested.
-                if self.hash:
-                    data = open(raw_filename, 'rb').read()
-                    assert get_hash(self.hash, data) == raw_info.hashes[self.hash]
+            # Validate if requested.
+            if self.validate_hash:
+                data = open(raw_filename, 'rb').read()
+                assert get_hash(self.validate_hash, data) == raw_info.hashes[self.validate_hash]
 
-    def _download_shard(self, shard: int, partition: Partition) -> int:
+    def _download_shard(self, shard_id: int) -> None:
         """Download the given shard.
 
         Args:
-            shard (int): Shard ID.
-            partition (Partition): Our rank and worker's partition of the dataset.
-
-        Returns:
-            int: Shard ID.
+            shard_id (int): Shard ID.
         """
-        assert shard in partition.shards
-        reader = self.shards[shard]
+        reader = self.shards[shard_id]
         for raw_info, zip_info in reader.file_pairs:
-            self._download_shard_part(shard, partition, raw_info, zip_info, reader.compression)
-        return shard
+            self._download_shard_part(shard_id, raw_info, zip_info, reader.compression)
 
-    def _download_shards_via_pool(self,
-                                  shards: List[int],
-                                  partition: Partition,
-                                  num_processes: Optional[int] = None) -> None:
-        """Download and load the given missing shards.
-
-        This is done in the main thread using a process pool.
+    def _download_or_await_shard(self, lock: FileLock, shard_states: NDArray[np.uint8],
+                                 shard_id: int) -> None:
+        """Either download the given shard or wait on its download.
 
         Args:
-            shards (List[int]): The missing shards to download.
-            partition (Partition): Our rank and worker's partition of the dataset.
-            num_processes (int, optional): Number of concurrent shard downloads (ie, size of the
-                process pool). If None, uses number of CPUs. Defaults to ``None``.
+            lock (FileLock): The lock protecting ``shard_states``.
+            shard_states (NDArray[np.uint8]): The download status of each shard, as an array in
+                shared memory.
+            shard_id (int): Shard ID.
         """
-        pool = Pool(num_processes)
-        download_shard = lambda shard: self._download_shard(shard, partition)
-        for shard in pool.imap_unordered(download_shard, shards):
-            self._load_shard(shard, partition)
-        with self._lock:
-            self._download_status = DownloadStatus.DONE
-
-    def _get_num_todo_samples(self) -> int:
-        """Get the number of available samples.
-
-        Returns:
-            int: Number of available samples.
-        """
-        min_size = None
-        with self._lock:
-            for todo_ids in self._epoch_to_todo_ids.values():
-                size = len(todo_ids)
-                if min_size is None or size < min_size:
-                    min_size = size
-        return min_size or 0
-
-    def _wait_until_few_todo_samples(self):
-        """Block until the samples are low enough to download another shard."""
-        if self.prefetch is None:
+        # First, the fast path: check the shared memory shard state without taking the lock. The
+        # shard states only ever go up, so if we're at the downloaded state, it's downloaded.
+        state = shard_states[shard_id]
+        if state == _ShardState.DOWNLOADED:
             return
-        while True:
-            if self._get_num_todo_samples() <= self.prefetch:
-                break
-            else:
-                sleep(0.25)
 
-    def _download_shards_via_loop(self, missing_shards: List[int], partition: Partition) -> None:
-        """Sequentially download and load the given missing shards.
-
-        This method is run in a background thread, which cannot use process pools because daemonic
-        threads can't have child processes. In any case, with every worker downloading shards at
-        once, process pool isn't necessary.
-
-        Args:
-            missing_shards (List[int]): The missing shards to download.
-            partition (Partition): This rank and worker's part of the dataset.
-        """
-        for shard in missing_shards:
-            self._wait_until_few_todo_samples()
-            try:
-                self._download_shard(shard, partition)
-                self._load_shard(shard, partition)
-            except Exception as e:
-                with self._lock:
-                    self._download_status = DownloadStatus.FAILED
-                    self._download_exception = e
-                return
-        with self._lock:
-            self._download_status = DownloadStatus.DONE
-
-    def download(self, num_processes: Optional[int] = None) -> None:
-        """Load all shards, downloading if not local (blocking).
-
-        Args:
-            num_processes (int, optional): Number of concurrent shard downloads (ie, size of the
-                process pool). If ``None``, uses number of CPUs. Defaults to ``None``.
-        """
-        partition = self.index.get_partition()
-        shards = self._preload(partition)
-        if shards:
-            self._download_shards_via_pool(shards, partition, num_processes)
-
-    def _start_downloading(self) -> bool:
-        """Load shards in a thread, returning whether done immediately.
-
-        Returns:
-            bool: Whether all the shards were already local (are now loaded).
-        """
-        partition = self.index.get_partition()
-        missing_shards = self._preload(partition)
-        if missing_shards:
-            Thread(target=self._download_shards_via_loop,
-                   args=(missing_shards, partition),
-                   daemon=True).start()
-        with self._lock:
-            return self._download_status == DownloadStatus.DONE
-
-    def _iter_ids_static(self) -> Iterator[int]:
-        """Get an iterator over all our sample IDs.
-
-        Returns:
-            Iterator[int]: Each sample ID.
-        """
-        ids = list(self._downloaded_ids)
-        if self.shuffle:
-            np.random.shuffle(ids)
-        yield from ids
-
-    def _iter_ids_dynamic(self) -> Iterator[int]:
-        """Get an iterator over all our sample IDs as they become downloaded.
-
-        If we are currently out of samples but not finished downloading the shards, blocks until it
-        has new samples.
-
-        Returns:
-            Iterator[int]: Each sample ID.
-        """
-        with self._lock:
-            epoch = self._next_epoch
-            self._next_epoch += 1
-            self._epoch_to_todo_ids[epoch] = todo_ids = list(self._downloaded_ids)
-        while True:
-            with self._lock:
-                if self._download_status == DownloadStatus.IN_PROGRESS:
-                    if todo_ids:
-                        yield todo_ids.pop()
-                        continue
-                elif self._download_status == DownloadStatus.DONE:
-                    if todo_ids:
-                        yield todo_ids.pop()
-                        continue
-                    else:
-                        del self._epoch_to_todo_ids[epoch]
-                        break
-                elif self._download_status == DownloadStatus.FAILED:
-                    raise self._download_exception
-                else:
-                    raise RuntimeError('Unexpected download status.')
-            sleep(0.25)
-
-    def _iter_ids(self) -> Iterator[int]:
-        """Get an iterator over all our sample IDs.
-
-        Returns:
-            Iterator[int]: Each sample ID.
-        """
-        if self._start_downloading():
-            yield from self._iter_ids_static()
+        # Shard is not necessarily downloaded, so check and update state with the lock.
+        lock.acquire()
+        state = shard_states[shard_id]
+        if state == _ShardState.UNKNOWN:
+            shard_states[shard_id] = _ShardState.DOWNLOADING
+            lock.release()
+            self._download_shard(shard_id)
+            with lock:
+                shard_states[shard_id] = _ShardState.DOWNLOADED
+        elif state == _ShardState.DOWNLOADING:
+            lock.release()
+            while shard_states[shard_id] != _ShardState.DOWNLOADED:
+                sleep(0.07)
+        elif state == _ShardState.DOWNLOADED:
+            lock.release()
         else:
-            yield from self._iter_ids_dynamic()
+            raise RuntimeError('Unknown shard state')
 
-    def __getitem__(self, idx: int) -> Any:
-        """Get sample by global index, blocking to load its shard if missing.
+    def _download_thread(self, epoch: int, sample_ids: NDArray[np.int64]) -> None:
+        """Download the relevant shards in the background while we are being iterated.
 
         Args:
-            idx (int): Sample index.
-
-        Returns:
-            Any: Sample data.
+            epoch (int): Which epoch. On noticing that a new epoch has started, we exit this thread
+                because a new one will soon be running, with different sample IDs.
+            sample_ids (NDArray[np.int64]): The samples to download the shards of.
         """
-        # Create lock in __getitem__ because we are prevented from putting it in __init__ because
-        # of DataLoader num_workers and fork/spawn semantics.
-        if not hasattr(self, '_lock'):
-            self._lock = RLock()
+        # Create or attach shard_states array.
+        name = f'{self._prefix}_shard_states'
+        size = len(self.shard_sizes) * np.uint8(0).nbytes
+        try:
+            shm = SharedMemory(name, True, size)
+        except:
+            shm = SharedMemory(name)
+        shard_states = np.ndarray(len(self.shard_sizes), buffer=shm.buf, dtype=np.uint8)
 
-        # Locate the shard and sample offset within that shard where the sample lives.
-        shard_idx, idx_in_shard = self.index.find_sample(idx)
+        filename = os.path.join('/tmp', 'mds', self._prefix, '_shard_states_filelock')
+        shard_states_lock = FileLock(filename)
 
-        # Load its shard if not loaded.
-        with self._lock:
-            if not self._has_shard[shard_idx]:
-                partition = self.index.get_partition()
-                self._download_shard(shard_idx, partition)
-                self._load_shard(shard_idx, partition)
+        # Download loop.
+        num_samples = len(sample_ids)
+        while True:
+            # If we've started a new epoch early (__iter__ was called again), exit this thread
+            # because there can only be one epoch at once.
+            if epoch != self.next_epoch - 1:
+                break
 
-        # Now that we have the shard, load the sample there.
-        shard = self.shards[shard_idx]
-        return shard[idx_in_shard]
+            # If we're out of samples this epoch, exit this thread because we are done downloading.
+            if self._download_index == num_samples:
+                break
 
-    def __iter__(self) -> Iterator[Any]:
+            # If we are requested to only pre-download so many samples, if we have as many or more
+            # downloaded already, we wait and check again later.
+            if self.predownload is not None:
+                samples_ahead = self._download_index - self._iter_index
+                if self.predownload <= samples_ahead:
+                    sleep(0.07)
+                    continue
+
+            # Download and decompress the shard for this sample, if not already done.
+            sample_id = sample_ids[self._download_index]
+            shard_id, _ = self.index.find_sample(sample_id)
+            self._download_or_await_shard(shard_states_lock, shard_states, shard_id)
+            self._download_index += 1
+
+    def __iter__(self) -> Iterator[Dict[str, Any]]:
         """Iterate over all the samples in our partition.
 
-        If not all samples have been downloaded yet, iterates over what it has while inserting the
-        remainder into the sequence behind the scenes as it progresses.
+        Returns:
+            Iterator[Dict[str, Any]]: Each sample.
+        """
+        # Load resume state and create the new training session.
+        presumed_epoch = self.next_epoch
+        epoch, old_sessions = self._resume(presumed_epoch)
+        cur_session, _ = self._create_cur_session(epoch)
+
+        self._barrier()
+
+        # Update the pre-incremented epoch counter.
+        world = World()
+        if world.is_local_leader:
+            self.next_epoch = epoch + 1
+
+        self._barrier()
+
+        # Get the samples for this worker to process.
+        sessions = old_sessions + [cur_session]
+        sample_ids = self._get_partition(epoch, sessions, world)
+
+        # Iterate while downloading.
+        self._iter_index = 0
+        self._download_index = 0
+        Thread(target=self._download_thread, args=(epoch, sample_ids)).run()
+        num_samples = len(sample_ids)
+        while self._iter_index < num_samples:
+            if self._iter_index < self._download_index:
+                cur_session[world.worker] += 1
+                sample_id = sample_ids[self._iter_index]
+                yield self[sample_id]
+                self._iter_index += 1
+                continue
+            sleep(0.07)
+
+        # Any code after the yields will never be reached by the Composer trainer.
+
+    def _all_gather_current_session(self, session: NDArray[np.int64]) -> None:
+        """All-gather the current session data.
+
+        This is done in order to checkpoint.
+
+        Args:
+            session (NDArray[np.int64]): The current session.
+        """
+        # Bail if we are not multi-node.
+        world = World()
+        if not world.is_multinode:
+            return
+
+        # Do the all_gather on the last session counts.
+        device = torch.device(f'cuda:{world.rank}')
+        source = torch.tensor(session, device=device)
+        dests = [
+            torch.empty(len(session), dtype=torch.int64, device=device)
+            for _ in range(world.num_ranks)
+        ]
+        dist.all_gather(dests, source)
+
+        # Each rank provides ground truth for its workers.
+        if world.is_local_leader:
+            dests = torch.stack(dests).cpu().numpy()  # Shape: (world size, total workers).
+            for rank in range(world.num_ranks):
+                rank_start = rank * world.workers_per_rank
+                rank_end = (rank + 1) * world.workers_per_rank
+                session[rank_start:rank_end] = dests[rank]
+
+        # Wait for local leaders to load session state from the other nodes.
+        dist.barrier()
+
+    def state_dict(self) -> Dict[str, Any]:
+        """Get a dict containing training state (called from non-worker process).
+
+        This is called on rank zero.
 
         Returns:
-            Iterator[Any]: Each sample.
+            Dict[str, Any]: The state.
         """
-        for idx in self._iter_ids():
-            yield self[idx]
+        # Attempt to load resume state, if it exists.
+        epoch, old_sessions = self._resume(self.next_epoch - 1)
+
+        # Get the current training session array, if we are currently training.
+        cur_session, _ = self._lookup_cur_session(epoch)
+
+        # Concatenate the sessions, synchronizing current session if we have one.
+        if cur_session is not None:
+            self._all_gather_current_session(cur_session)
+            sessions = old_sessions + [cur_session]
+        else:
+            sessions = old_sessions
+
+        return {
+            'epoch': epoch,
+            'sessions': [x.tolist() for x in sessions],
+        }
+
+    def load_state_dict(self, obj: Dict[str, Any]) -> None:
+        """Load a dict containing training state (called from non-worker process).
+
+        This is called on each copy of the dataset when resuming.
+
+        Args:
+            obj (Dict[str, Any]): The state.
+        """
+        # Set the number of the next epoch.
+        self.next_epoch = obj['epoch']
+
+        # Save the resume state (old sessions).
+        name = f'{self._prefix}_resume'
+        data = json.dumps(obj, sort_keys=True).encode('utf-8')
+        try:
+            self._resume_shm = SharedMemory(name, True, len(data))
+            self._resume_shm.buf[:] = data
+        except:
+            self._resume_shm = SharedMemory(name)
+            assert len(self._resume_shm.buf) == len(data)
