+++ streaming/vision/__init__.py
@@ -5,7 +5,6 @@
 
 from streaming.vision.ade20k import ADE20K as ADE20K
 from streaming.vision.cifar10 import CIFAR10 as CIFAR10
-from streaming.vision.cifar10 import LocalResumableCIFAR10 as LocalResumableCIFAR10
 from streaming.vision.coco import COCO as COCO
 from streaming.vision.imagenet import ImageNet as ImageNet
